Fix file types
