+++ .stylelintignore
@@ -2,4 +2,4 @@ build
 build-style
 node_modules
 packages/stylelint-config/test
-packages/theme/src/stylelint-plugins/test
+packages/theme/src/stylelint-plugins/test/fixtures
