+++ packages/theme/src/stylelint-plugins/test/.stylelintignore
@@ -1,0 +1,1 @@
+# This file needs to exist so test files have a way to unignore the files in this folder.
